✨ Add csv files, fix cuda bug for bigger heatmaps (blocks per grid)

+++ heatmap-cuda/heatmap.cu
@@ -76,7 +76,7 @@ int Heatmap::getHeight()
     return height;
 }
 
-int Heatmap::getSize()
+long Heatmap::getSize()
 {
     return width * height;
 }
